split seeder from rand_seed, make rand_seed a sub-module of it
