+++ src/generator/seeder/rand_seed.rs
@@ -1,21 +1,6 @@
+use super::{SeedValue, Seeder};
 use alloy::primitives::{keccak256, U256};
 use rand::Rng;
-
-pub trait Seeder {
-    fn seed_values(
-        &self,
-        amount: usize,
-        min: Option<U256>,
-        max: Option<U256>,
-    ) -> Box<Vec<impl SeedValue>>;
-}
-
-pub trait SeedValue {
-    fn as_bytes(&self) -> &[u8];
-    fn as_u64(&self) -> u64;
-    fn as_u128(&self) -> u128;
-    fn as_u256(&self) -> U256;
-}
 
 /// Default seed generator, using a random 32-byte seed.
 #[derive(Debug, Clone)]
